refactor(trainer.plugins): Add plugins missing and updates for deprecated methods
