+++ trainer/plugins/plugin.py
@@ -1,0 +1,9 @@
+class Plugin(object):
+
+    def __init__(self, interval=None):
+        if interval is None:
+            interval = []
+        self.trigger_interval = interval
+
+    def register(self, trainer):
+        raise NotImplementedError
